fix: stop silently coercing non-numeric IAM_AUTH_CREDENTIALS_EXPIRY_BUFFER
The (int) cast in the config closure turned 'not-a-number' into 0 before
the validator could see it, silently disabling the buffer. Drop the cast
and let the validator/coercion helpers handle non-numeric values.

+++ config/iam-auth.php
@@ -80,7 +80,7 @@ return [
     |
     */
 
-    'credentials_expiry_buffer' => (int) env('IAM_AUTH_CREDENTIALS_EXPIRY_BUFFER', 10),
+    'credentials_expiry_buffer' => env('IAM_AUTH_CREDENTIALS_EXPIRY_BUFFER', 10),
 
     /*
     |--------------------------------------------------------------------------
